+++ modules/modules_manager/builtin/cpp/semverBuiltin.cpp
@@ -1,0 +1,55 @@
+//=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+//
+// Alternatively, you can redistribute it and/or
+// modify it under the terms of the GNU General Public License as
+// published by the Free Software Foundation; either version 2 of
+// the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public
+// License along with this program. If not, see <http://www.gnu.org/licenses/>.
+// LICENCE_BLOCK_END
+//=============================================================================
+#include "semverBuiltin.hpp"
+#include "Error.hpp"
+#include "CompareVersions.hpp"
+#include "characters_encoding.hpp"
+//=============================================================================
+using namespace Nelson;
+//=============================================================================
+ArrayOfVector
+Nelson::ModulesManagerGateway::semverBuiltin(
+    Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
+{
+    ArrayOfVector retval;
+    if (argIn.size() != 2) {
+        Error(ERROR_WRONG_NUMBERS_INPUT_ARGS);
+    }
+    if (nLhs > 1) {
+        Error(ERROR_WRONG_NUMBERS_OUTPUT_ARGS);
+    }
+    std::string param1 = argIn[0].getContentAsCString();
+    std::string param2 = argIn[1].getContentAsCString();
+    std::string errorMessage;
+    int value = CompareVersions(param1, param2, errorMessage);
+    if (!errorMessage.empty()) {
+        Error(_W("Cannot compare versions: ") + utf8_to_wstring(errorMessage));
+    } else {
+        retval.push_back(ArrayOf::doubleConstructor(value));
+    }
+    return retval;
+}
+//=============================================================================